Moved all files related to matching engine to one folder. They should be rearranged in the future.
