+++ GomokuServer/src/GomokuServer.Api/MatchingEngine/GameOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace GomokuServer.Api;
+﻿namespace GomokuServer.Api.MatchingEngine;
 
 [TranspilationSource]
 public record GameOptions(int BoardSize, TimeControlDto TimeControl, bool Anonymous);
