+++ GomokuServer/src/GomokuServer.Api/MatchingEngine/IMatchingEngine.cs
@@ -1,4 +1,4 @@
-﻿namespace GomokuServer.Api;
+﻿namespace GomokuServer.Api.MatchingEngine;
 
 public interface IMatchingEngine
 {
