+++ GomokuServer/src/GomokuServer.Api/MatchingEngine/QuickPairingGameOptionsVariants.cs
@@ -1,4 +1,4 @@
-﻿namespace GomokuServer.Api;
+﻿namespace GomokuServer.Api.MatchingEngine;
 
 public static class QuickPairingGameOptionsVariants
 {
